requirements.txt
loguru==0.7.2
numpy==2.1.2
Pillow==10.4.0
pydantic==2.9.2
timm==1.0.9
torch==2.3.1
torchvision==0.18.1
